+++ base-compat-batteries/src/Text/Read/Lex/Compat.hs
@@ -1,0 +1,6 @@
+{-# LANGUAGE CPP, NoImplicitPrelude, PackageImports #-}
+module Text.Read.Lex.Compat (
+  module Base
+) where
+
+import "base-compat" Text.Read.Lex.Compat as Base